[DPR2-71] Extract mapping from service
